Expanded TextureDescs, not sure how I should approch the arrayerMaster and ArraySlave items Rearranged includes, to get rid of the damn redeffinition error added forward declaration of DemoMeister and moved the Extern variable of DemoMeister singleton to Util.h
Now it should be available everywhere.
I belive however that it should not be referenced in the scenegraph components, but oh well.
I have plenty of entry points, I really could try make some of the components dependency neutral.

+++ Demoscene/Engine/AssetMapClasses/map.cpp
@@ -108,7 +108,7 @@ GLint MapAsset::LoadMaps(TextureDesc  *ToCreate)
 	
 	//I sorely need a more durable and conscice ID, what more it should store a direct access index.
     //http://en.cppreference.com/w/cpp/types/ptrdiff_t
-	NS_ENG::MapAsset::classMapList.back()->Map_MapID = NS_ENG::MapAsset::classMapList.size();
+	NS_ENG::MapAsset::classMapList.back()->Map_MapID = (GLint)NS_ENG::MapAsset::classMapList.size();
 	//NS_ENG::MapAsset::classMapList.push_back(tmp_TexMap);
 	//NS_ENG::MapAsset::classMapList.back().Map_MapID = NS_ENG::MapAsset::classMapList.size();
 	//NS_ENG::MapAsset::classMapList.back()->MapIter = NS_ENG::MapAsset::classMapList.size();
